changes following review feedback

+++ livingatlas/pipelines/src/main/java/au/org/ala/pipelines/vocabulary/CentrePoints.java
@@ -174,7 +174,7 @@ public class CentrePoints {
       for (int i = 0; i < decimalPlaces; i++) {
         x = x * 10;
       }
-      return ((double) (Math.round(number * x))) / x;
+      return Math.round(number * x) / x;
     } else {
       return Math.round(number);
     }
